seeder作成中　42S22エラー
